refactor : OutputView 이름 변경
